include camera model and shorten exif stats
This adds a line to show the camera model and make, if available, and
reorders and cleans up the EXIF parameters extracted so they are more
readable for photographers (who are the ones who care about this
anyways).

Note that, as a photographer, I know what "Fstop" mean, but I would
expect "exposure" or the common f/X.Y notation. The "Focal" label
means nothing to me: "Focal length" would mean something, but it's
long - and what are the units? Just using "mm" here clarifies any
possible misunderstanding.

I have found this format to be way more readable. The following:

    ISO: 3200, Focal: 50, Exposure: 1/40, Fstop: 4.0
    Date: dimanche, 31. décembre 2017

Becomes:

    ISO 3200, 1/40s f/4.0, 50mm
    FUJIFILM X-T2
    dimanche, 31. décembre 2017

Now: dropping the explicit labels may be controversial, but I would
argue this is what markup is for. We could wrap those elements in
named <span> or <acronym> elements to provide more information through
mouse-over or CSS.